django/requirements.txt
Django==2.1.3
PyMySQL==0.9.2
mysqlclient==1.4.1
django-compressor==2.2
pygraphviz==1.5
ERAlchemy==1.2.10
PyYAML==3.12
argon2-cffi==19.1.0
pyminizip==0.2.4
gunicorn==19.9.0
django-environ==0.4.5
django-widget-tweaks==1.4.3
rollbar~=0.14.7
django-libsass~=0.7
djangorestframework~=3.10.3
django-filter~=2.2.0
django-webpack-loader~=0.6.0
pandas~=0.25.1
